feat (ci): Support manually running CI tests (#1556)
# Motivation
It is useful to be able to trigger CI checks manually.  For example, the backend test is triggered only on changes to certain files, so it is not easy to check that it works on main.  With "workflow dispatch" enabled, the backend tests can be run on main with the click of a button.

# Changes
- Add "workflow_dispatch" as a trigger for those GitHub CI jobs that don't have it yet.

# Tests
- This needs to be merged into main before the "workflow dispatch" functionality can be exercised.