fix (ci): Env vars are not supported for the runner (#1558)
# Motivation
The backend test [fails on main](https://github.com/dfinity/oisy-wallet/actions/runs/9658380600) with "invalid workflow file".  In particular Ci doesn't like the use of env in the `runs-on` clause.  The backend tests [have been failing without exception for a couple of months](https://github.com/dfinity/oisy-wallet/actions/workflows/backend-tests.yml).

I see no evidence of env variables being supported for runs-on in the CI docs or functioning examples in GitHub.  `var` can be used but that is different.

While it would be nice to factor out the runner, something simple that works is better than something fancy that doesn't. (yet!)

# Changes
- Specify the backend test runner manually.

# Tests
- See the backend tests succeeding with this change: https://github.com/dfinity/oisy-wallet/actions/runs/9658643523
  - Note: This run was triggered manually.