feat(devops): Add Init Args (#1773)
Motivation
To validate credentials Oisy backend will need some information about the environment. Such as the canister ids of II and the POUH issuer.

In this PR, I add new fields to InitArg and pass them on deploy.

Changes
Add fields to InitArg.
Gather the environment information in deploy.backend and pass it as init arguments.
Update candid file with new fields and type.
Update the config on post_upgrade and not only on init.
Tests
I tested by deploying locally the backend to an already deployed canister and it worked.