chore(frontend): beta configuration (#1856)
# Motivation

We were requested to deploy a new frontend under `beta.oisy.com` that
points to production. For this purpose, the frontend will communicate
with the same backend and canisters as production, and we will use a
derivation origin to sign in with `oisy.com`.

# Notes

We will use two separate environments because, while they are the same
right now, I would not be surprised if "beta" has to be tweaked with
different environment properties. Moreover, it is more straightforward
to do so.

To make `beta.oisy.com` point to the `oisy.com` backend, we will
duplicate the backend ID in `canister_ids.json`.

# Changes

We reviewed the changes offline with @AntonioVentilii-DFINITY

---------

Signed-off-by: David Dal Busco <david.dalbusco@outlook.com>