feat(frontend): Add service to load user profile (#1847)
# Motivation

We want to use the profile stored in the backend in the frontend.

In this PR, a new service is introduced `loadUserProfile` which loads
gets or creates a profile and stores it in the store. It also performs
an update call to store the certified data in the background.

# Changes

* New api function `getUserProfile` and `createUserProfile`.
* New error `UserProfileNotFoundError`.
* New service `loadUserProfile`.

# Tests

* Unit tests for `loadUserProfile`.
* I also tested triggering the service when the user visited settings
and it worked as expected.