feat(frontend): new send modal for all tokens (#1877)
# Motivation

We include the possibility to send tokens from the main page. To do
that, we first create a modal that "merge" both flows for IC and ETH
tokens.
The modal will display only the enabled tokens that have non-null
balance.

# Changes

- Create new component `SendTokensList` with the desired list of tokens:
   - PR #1887 
- Create `SendModal` with the usage of both IC and ETH send modals:
  - PR #1886 
  - PR #1907 
- Create `Send` button.

# Tests

I sent out ICP and ETH on local replica and it worked as expected, as
usual.