refactor(backend): Add UserProfileModel to interact with user profiles in stable memory (#1932)
# Motivation

Refactor the code so that the different maps managing user profiles are
not mismanaged and we introduce bugs and inconsistent data.

This was requested in
https://github.com/dfinity/oisy-wallet/pull/1827/files#r1695268288

# Changes

* Create a new struct UserProfileModel and implement it. This entity
will have the responsibility to interact with the stable memory to
retrieve and store user profiles.
* Use the new struct in the user_profile module instead of expecting the
stable memory structures.

# Tests

Not yet. I want to check that the implementation is the requested one.