feat(backend): Controllers are allowed (#1950)
# Motivation
Some `backend` canister APIs are gated to authorized maintainers. The
list of maintainers is kept in a list of `allowed_users`. Controllers of
this canister are also maintainers and, as discussed with the oisy team,
it seems reasonable to consider controllers to be allowed to call these
endpoints.

# Changes
* Allow canister controllers to call restricted endpoints.
* Update the pocket-ic setup to set controllers.

# Tests
The `config()` API is has tests that only allowed users can call the
API. I have updated this to check that controllers are able to call
`config()`.