feat(backend): Private stats API (#1980)
# Motivation
When migrating user data, it would be useful to be able to see how many
users and other data is in the old and the new backend canister.

Data such as the number of users is not currently public, as far as I
can tell, so I have kept this data restricted to "allowed" users.

# Changes
- Add a private `stats()` query endpoint to support checking migration
progress.

# Tests
- `PocketIc` tests are included to ensure that the stats endpoint gives
expected values and is accessible only to authorized users.