feat(backend): Support multiple backends in tests (#1982)
# Motivation
The current tests each create a test environment consisting of the tuple
`(PocketIc, Principal)` where the Principal is the backend canister ID.
This test environment is passed around as an argument to various
utilities. This approach has issues as soon as there are two backend
canisters. E.g. ownership of the PocketIc instance prevents us from
having two such tuples. We need a more flexible approach.

# Changes
- Create a struct that can take the place of the existing tuples.
- That struct uses an `Arc` to allow several structs to share a single
`PocketIc`.
- Copy (and adapt) utilities to these new structs.

Note: The bulk update of tests is NOT done in this PR in order to keep
the change small and readable.

# Tests
These structs are as yet unused, due to breaking changes into small PRs.
The structs here are used in new tests in a feature branch; the test
structures work there and are ergonomic. A review of existing tests
suggests that adapting them to use these structures instead of tuples
should be fairly straightforward.