feat(frontend): gather some anonymous analytics about transactions duration (#1996)
# Motivation

We discussed with @DSharifi if there was an easy way to gather analytics
and insights about the duration of processes as perceived by users. As a
result, Daniel had the idea to persist the duration of function
executions as anonymous metadata in the events we save in Juno
analytics. Basically could potentially be a quick win. This PR provides
some of the first metrics.

# Changes

- Add two new functions in the analytics modules to initiate events and
ultimately submit those once a function is executed
- Track those kind of events for ETH, IC, ICRC and ERC20 transactions

Signed-off-by: David Dal Busco <david.dalbusco@outlook.com>