fix(frontend): avoid multiple balance refresh periodically (#2029)
# Motivation

As spotted by @peterpeterparker , the ETH/ERC20 balance was being
refreshed periodically and that could cause incoherent and asynchronous
information between the balance and the transactions.

The cause is most likely a cascade refresh caused by the periodic
refresh of `exchanges` store:

`exchanges --> combinedDerivedSortedTokens --> networkTokens -->
enabledNetworkTokens --> enabledErc20NetworkTokens --> LoaderBalances`

This PR tries to solve that issue removing the exchanges dependency from
`enabledErc20NetworkTokens` dependencies.

To do that, we move `combinedDerivedSortedTokens`, so that the sorting
happens after having filtered for `network` and `enabled`. This has the
additional effect of potentially reducing the computational work for
sorting.