refactor(frontend): use common initCertifiedSetterStore for addressStore (#2012)
# Motivation

Since we want to have a series of derived addresses corresponding to
their token (and right now we have just ETH), we change the
`addressStore` to use `initCertifiedSetterStore`, that is a common store
in the code basically returning a record, with certified data. We
already use it for example for `balanceStore`.

# Tests

No changes in local replica.