feat(devops): Add signer canister (#2045)
# Motivation
The backend will be split into a signer and a user data canister. The
first step in this process, according to [the
plan](https://www.notion.so/dfinityorg/Oisy-frontend-uses-separate-backend-and-signer-6a84a876f6ec434891eafcb56756937c),
is to have separate canister names, even though they initially refer to
the same canister.

# Changes
- Add a `signer` canister to `dfx.json` and `canister_ids.json`.

# Tests
- Existing CI should pass.
- Reviewers should manually verify that, at the time of writing, the
canister IDs are identical for the `backend` and `signer`.