fix(frontend): ICP Erc20 to native ICP destination account identifier (#2048)
# Motivation

To enable the conversion of "ICP ERC20" tokens to native ICP tokens,
users should be able to input an "old" ICP account identifier in
hexadecimal format. Previously, this function worked as expected.
However, due to an assertion introduced in early 2024, the destination
field in the ETH "Send" wizard now requires a properly formatted ETH
address. To accommodate this specific use case, we need to modify the
validation process to recognize and accept ICP account identifiers as
valid inputs in the destination field.

# Notes

There might be some more abstracted ways to solve this but, to prevent
issues, I rather liked to bit a bit verbose given that it is a really
particular use case.

# Changes

- Improve `isInvalidDestination` to throw if the address is an invalid
account identifier on "ICP Erc20"
- Do not map to Eth address prefixed with `0x` on send
- Do not map to Eth address prefixed with `0x` when calculating fee