feat(backend): User data migration (#1967)
# Motivation
We would like to migrate user data to a new canister. The groundwork for
the migration code is already in place. Now we need the actual migration
code.

# Changes
- Populate the skeleton code for migrating data.
- Populate the skeleton e2e test.

# Tests
A `pocket-ic` test is included that exercises the migration from end to
end. The main difference from production code is that the `pocket-ic`
code steps explicitly whereas normally the migration is driven by a
timer.