feat(devops): Staging backend has moved (#2093)
# Motivation
The staging backend has moved to `new_backend`:
- The user data is now in the new backend, signing is unavailable from
the new backend.
- The old backend is now the signer, it no longer handles user data
requests.

# Changes
- Update entries in `canister_ids.json`
- Fix the II login issue: The signer API needs to use the signer
backend.

# Tests
Signer has signing enabled, user data disabled:
```
$ dfx canister call --network staging signer config
WARN: The default identity is not stored securely. Do not use it to control a lot of cycles/ICP. Create a new identity with `dfx identity new` and use it in mainnet-facing commands with the `--identity` flag
(
  record {
    api = opt record {
      user_data = variant { Disabled };
      threshold_key = variant { Enabled };
    };
```
Backend is the other way around:
```
$ dfx canister call --network staging backend config
WARN: The default identity is not stored securely. Do not use it to control a lot of cycles/ICP. Create a new identity with `dfx identity new` and use it in mainnet-facing commands with the `--identity` flag
(
  record {
    api = opt record {
      user_data = variant { Enabled };
      threshold_key = variant { Disabled };
    };
```
Backend has user data:
```
$ dfx canister call --network staging backend stats
WARN: The default identity is not stored securely. Do not use it to control a lot of cycles/ICP. Create a new identity with `dfx identity new` and use it in mainnet-facing commands with the `--identity` flag
(
  record {
    user_profile_count = 4 : nat64;
    custom_token_count = 8 : nat64;
    user_timestamps_count = 4 : nat64;
    user_token_count = 10 : nat64;
  },
)
```