style(frontend): animate token cards when tokens list changes (#1781)
# Motivation

We would like a smooth animation when the cards in the token list change
order. This could happen since we have now a dynamic sorting based on
market data too. Furthermore, we disable clicking on the cards while
they move around.

Note: an element that uses the animate directive must be the immediate
child of a keyed each block.

The below recording shows what happens when the list is randomly
shuffled.


https://github.com/user-attachments/assets/f913a03b-f515-4c67-ba05-2031fc917332