fix(frontend): pending ck conversions displayed under incorrect token (#2103)
# Motivation

It was reported that pending ck conversions were displayed under the
wrong token. e.g. one user converted LINK to ckLINK and then browsed
ckUSDC and the frontend was displaying "Pending USDC to ckUSDC". This
seems to happens because the pending transactions resolved with logs are
not filtered for the selected token.

To solve it, we refine the logs topic including the ERC20 contract
address among the parameters (encoding the token address).

We can see a practical example
[here](https://etherscan.io/tx/0xddbf4aebe292c3d3b33240496db12a1a5622c4edb63017ebcd12ffa32fa5b924#eventlog).

<img width="1302" alt="Screenshot 2024-08-20 at 16 14 30"
src="https://github.com/user-attachments/assets/4aa209b4-02f1-4cdc-b1a5-c0be0321020f">


# Note

This is the second attempt to solve the issue, after the PR #2069, that
was ultimately reverted.

---------

Co-authored-by: David Dal Busco <david.dalbusco@dfinity.org>