feat(frontend): add balance to TokenUi (#2150)
# Motivation

The balance store is called a few times in the code, related to derived
store `combinedDerivedEnabledNetworkTokensUi`. So, to simplify the code,
we add the prop `balance` to type `TokenUi`.
After this PR, we will make use of the new prop throughout the code.

# Changes

- Add `balance` to `TokenUi`.
- Calculate balance in store `combinedDerivedEnabledNetworkTokensUi`.

# Tests

No changes in local replica, just more loading time (that should be
compensated once we use the new prop).