chore(frontend): adjust Ethereum estimated fee label (#2189)
# Motivation

In the code, apart from conservative hard-coded numbers, we calculate
the ck-conversion fees related to Ethereum fetched from the minter using
the `eip_1559_transaction_price` call.
The `max_transaction_fee` fetched with call `eip_1559_transaction_price`
from the minter is not real-time: it is re-calculated everytime there is
a new withdrawal request to the minter. The same call provides the last
update timestamp.

So, in this PR we remove the `30 seconds` proxy string.