feat(frontend): Set up bitcoin listener (#2204)
# Motivation

There was an error in the console when the Bitcoin token was present in
the list of tokens.

The error came because the EthListener was used in the Bitcoin token.

This PR provides the component that must implement the listener for
bitcoin. But it doesn't implement it yet.

# Changes

* New empty BitcoinListener component.
* Check also for bitcoin network in Listener project.

# Tests

Tested locally that the error is not raised anymore.