feat(frontend): split address store [revert PR #2012] (#2207)
# Motivation

As discussed in PR #2067 , we would actually need to split the
`addressStore` into separate stores for each address. That was how it
was done before PR #2012 . So, as first step we partially revert that
PR, creating a specific `initAddressStore` function and using it for two
separate stores for BTC mainnet and ETH.

They will be used in another PR, where we substitute the usage of
`addressStore`.