feat(frontend): enabled IC network tokens (#2213)
# Motivation

To load balances for all enabled IC tokens, it will be useful to have a
specific derived store, similar to what we have for ERC20 tokens.

Note that we are not using `['icp', 'icrc'].includes(....` because this
is type checked, i.e. it would give an error if there was a mistype in
the right-hand side of the comparison that does not match one of the
possible values of TokenStandard.