feat(frontend): add support for WETH (#2212)
# Motivation

We include support for canonical Wrapped ETH (WETH). Official GitHub
Repo is [here](https://github.com/gnosis/canonical-weth).

For the logo, I found some difficulties, since I cannot find a central
place, documentation or source code about it.

Furthermore, the logo presented by
[Etherscan](https://etherscan.io/token/0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2),
[CoinMarketCap](https://coinmarketcap.com/currencies/weth/) and
[Metamask](https://github.com/MetaMask/contract-metadata/blob/master/images/weth.svg)
is different from the one shown by
[Uniswap](https://github.com/Uniswap/assets/blob/master/blockchains/ethereum/assets/0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2/logo.png)
and
[TrustWallet](https://github.com/trustwallet/assets/blob/master/blockchains/ethereum/assets/0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2/logo.png).

Since the last update from Uniswap and TrustWallet is 5 years ago and
Metamask is 2 years ago, I will use the one Metamask as source. I was
unable to find another source.

NOTE: in other repos of the same organization
[gnosis](https://github.com/gnosis), the logo is the one discarded,
however it seems that it was not updated in 4 years.
[Here](https://github.com/gnosis/corona-information-markets/blob/master/app/src/components/common/icons/currencies/WEthIcon.tsx)
and
[here](https://github.com/gnosis/conditional-tokens-explorer/blob/develop/src/components/common/TokenIcon/img/WEthIcon.tsx).