fix(frontend): wallets loader for all enabled ICRC tokens (#2210)
# Motivation

The component `IcListener` (that listens to each IC token balance, among
other things) is mounted only when the token is already present in the
list showed by `TokensSignedIn`.

That means that if the list is pre-mounted hiding zero-balances, the IC
tokens are not present in the list (since the balance by default is null
when the list is mounted) and consequently none of the `IcListener` is
started, keeping the balance of IC tokens null.

# Changes

- Create derived store for the simple list of enabled ICRC tokens
(including ICP). That is similar to what it is done with ERC20 tokens.
- Create **IcLoaderWallets** component: it replicates what
**IcListener** was doing, but in the entire list of tokens. It creates a
worker for each enabled IC token, and stop/remove the ones that are
disabled (or not belong to the network). Furthermore, we avoid blocking
the flow if one of the workers gives an issue. That would maintain the
same behavior as before.
- Use `IcLoaderWallets` in `Loaders`.
- Remove `IcListener` from `Listener`.

# Tests

In local replica the function is called correctly, and for the correct
amount of times. I enabled and disabled a token to test it too (see
log).

<img width="1276" alt="Screenshot 2024-09-03 at 10 38 16"
src="https://github.com/user-attachments/assets/8aba62a1-dd9e-406b-bf7f-ae0b6ed60329">