feat(frontend): track analytics when IC send completed (#2227)
# Motivation

There is a 2s timeout after an IC send transactions is complemented to
optimistically fetch the transactions. Given that we are interested to
get to know how long does a send least, it's not really correct to count
that timeout within the overall send. Therefore we can track the
anonymous metrics when the send is over.

# Changes

- Pass the analytics as callback for `sendIc` and collect analytics when
send is done.