feat(lint): add lint rule to sort svelte imports (#2231)
# Motivation

It seems like a good point to start: we would like to introduce the
[sorting import
rule](https://github.com/import-js/eslint-plugin-import/blob/HEAD/docs/rules/order.md)
for svelte component. However, done it drastically for the whole project
requires a few hundreds of fixing. So, we start doing it folder by
folder.

Gradually, we will include the entire project, at which point we move
the rule in the main `rules` prop.