refactor(frontend): move invalidIcpAddress to lib folder (#2232)
# Motivation

Since it is being used both by ETH and ICP components/functions, we move
`invalidIcpAddress` to folder `lib`. This is the same practice that
other functions in the same module have: for example `isEthAddress` is
used both by ETH and ICP modules.


As side effect of this, we have one less possible loop dependency inside
`icp` folder.