feat(frontend): update network-tokens to enabled tokens balance (#2233)
# Motivation

Step 3 of updating the chain selector with aggregate USD balance values:
adjust `network-tokens.derived` to load enabled tokens balance
regardless if a network selected by user or not. We discussed with
@AntonioVentilii-DFINITY the performance aspect of this change, and
since it will be basically the amount of data loaded (and hence time
needed to load it) as if we were in Chain Fusion, it should not affect
the user experience and performance.