fix(frontend): token balance could be null or undefined (#2281)
# Motivation

There is a difference between an `undefined` balance and a `null`
balance. The balance is `undefined` if the balance store is not
initiated or the specific balance loader for the token is not initiated.
But, if the balance loader was initiated at some point, it will either
contain data or be `null`, but not `undefined`.

We distinguish the two cases to make it easier to understand when a
Token could still need to wait for the balance loader to exists, or if
the balance loader is simply not working
(disconnected/stopped/error/etc).