ci(e2e): re-use oisy backend (#2205)
# Motivation

The idea is to cache backend WASM and expose it as an artifact in order
to fasten e2e and other pipelines. The cache will be invalidated on
backend files change. In this PR, I'm using the generated
`backend.wasm.gz` in Dockerfile.e2e by updating `dfx.json` to use
`custom` build instead of `rust`. The time needed to run FE e2e tests is
now decreased from ~22mins to ~8mins.