feat(devops): Add scripts to set up Bitcoin local node (#2228)
# Motivation

We want to be able to test that the development of the Bitcoin
integration works as expected.

In order to do that, we need a local Bitcoin node.

This PR provides two main scripts to set up the local node and start dfx
with Bitcoin. As well as some other utilities to be used while testing.

# Changes

* Add a couple of entries to `.gitignore`.
* Add a script to set up the local Bitcoin node.
* Add a script to start dfx with Bitcoin.
* Add a script to mine tokens.
* Add a section in Hacking.md to explain how to use the scripts.

# Tests

I and Denys tested locally that the scripts worked as expected.