refactor(frontend): common service to auto-load token (#2340)
# Motivation

The functions that auto-load tokens (both for ERC20 and ICRC) seem to
have the same patterns. This PR aims on reducing code repetition
creating a common function to be re-used
A few logic changes happens to achieve that:
- When the common function finds the counter-token, it will not check
the `standard`, since it is implied that the find function will return
the token that it is proper.
- We provide the sendToken directly casted as the expected type. That is
not necessary an issue, since the first check is specific on the
`standard` of such a token.