feat(frontend): listen to ETH mined transactions instead of pending ones [GIX-2909] (#2346)
# Motivation

Listening to pending transactions will only provide pending transactions
in the mempool in a prompt manner. That is a limitative since we would
like to display the user's ETH (and SepoliaETH) transactions as soon as
they are not-pending (excluding non-approved transactions).

This PR changes the transaction listener to listen to mined transactions
instead of pending ones.

Furthermore, we add a `slide` transition for the list of transactions.