feat(frontend): Onramper util to create the widget source link [GIX-2854] (#2488)
# Motivation

We create an util to generate the source link to provide to the Onramper
widget, given a set of parameters. Documentation can be found
[here](https://docs.onramper.com/docs/supported-widget-parameters).

For now we define a small set of FIAT currencies, just for the scope of
testing. The list can be expanded or improved after.

# Changes

- Create ENV variables for Onramper API keys.
- Add types for params of the new util.
- Create utils to build the source code and concatenate all params into
a single string.
- Create tests.

# Tests

Create some tests to verify the correct widget link: the links were
manually tested in the [Onramper
sandbox](https://docs.onramper.com/docs/customise-the-ux) and simply
opening them in the browser with the test API key.