chore(frontend): Update ckbtc dep in package.lock and package.json (#2534)
# Motivation

I realized that the BitcoinNetwork from ckbtc didn't have the latest
changes. Specifically, it didn't include "regtest" in the Bitcoin
networks.

The problem was that even though the package.json indicated a next
version, package-lock was still pointing to 3.0.0

In this PR I updated both package.json 

# Changes

* Change the package.json and package.lock.json with `npm rm
@dfinity/ckbtc && npm i @dfinity/ckbtc@next`.

# Tests

Code still works.