feat(backend): Add endpoint to select utxos and fee for bitcoin transaction. (#2562)
# Motivation

To make a bitcoin transaction utxos and a fee must be selected.

In this PR, I introduce an endpoint that calculates the utxos for a
transaction with a specific amount and then the fee related to it.

# Changes

## Manual changes

* New bitcoin_api module with exposed functions: `get_all_utxos` and
`get_fee_per_byte`.
* New bitcoin_utils module with expose functions: `estimate_fee` and
`utxos_selection`. These utils are based on the ckBTC minter's code.
* New endpoint `btc_select_user_utxos_fee` that uses helpers to
calculate and return utxos and fees.
* New endpoint types.

## Automatic changes

* Changes in `backend.did` with `./scripts.did.sh`.
* Changes in `src/declarations/backend/*` with `npm run generate`.

# Tests

* Unit tests for the bitcoin_utils.
* Dummy integration test for the new endpoint. Pocket-ic doesn't support
(yet) bitcoin.